feat: add mobile slider; refact: new js modules, delete extra props
