+++ src/blocks/bikes/bikes.scss
@@ -14,6 +14,8 @@
 
   @media screen and (max-width: 767px) {
     padding: 40px $paddingHorizontalSectionLargeMobile;
+    overflow: hidden;
+    position: relative;
   }
 
   @media screen and (max-width: 425px) {
@@ -22,7 +24,7 @@
     &::before {
       content: '';
       display: block;
-      border: 1px solid #e3e3e3;
+      border-top: 1px solid #e3e3e3;
       margin-bottom: 40px;
       width: 100%;
     }
