+++ src/blocks/card/__caption/_type/card__caption_type_inside.scss
@@ -6,6 +6,5 @@
     font-size: 14px;
     line-height: 18px;
     font-weight: $semiBold;
-    text-align: center;
   }
 }
