+++ src/blocks/card/card.scss
@@ -2,5 +2,8 @@
   width: 100%;
   box-sizing: border-box;
   max-width: 380px;
-  position: relative;
+
+  @media screen and (max-width: 767px) {
+    min-width: 100%;
+  }
 }
