+++ src/blocks/cards-container/cards-container.scss
@@ -16,5 +16,8 @@
   @media screen and (max-width: 767px) {
     padding-top: 20px;
     column-gap: 0;
+    position: relative;
+    left: 0;
+    transition: left 1s ease;
   }
 }
