+++ src/blocks/description/description.scss
@@ -23,7 +23,7 @@
     &::before {
       content: '';
       display: block;
-      border: 1px solid #e3e3e3;
+      border-top: 1px solid #e3e3e3;
       margin-bottom: 40px;
       width: 100%;
     }
