+++ src/blocks/intro/__bike-card/intro__bike-card.scss
@@ -8,7 +8,7 @@
   border-radius: 20px;
   background-color: $backgroundWhiteTheme;
   @include flex(row, center, center);
-  box-shadow: -20px 60px 220px rgba(0, 0, 0, 0.44);
+  filter: drop-shadow(0px 20px 140px rgba(0, 0, 0, 0.44));
 
   @media screen and (max-width: 768px) {
     max-width: 320px;
