+++ src/blocks/menu/__button/menu__button.scss
@@ -10,44 +10,8 @@
   @media screen and (max-width: 600px) {
     display: block;
   }
-}
-
-.menu__button-icon {
-  display: none;
-
-  @media screen and (max-width: 600px) {
-    display: block;
-    width: 22px;
-    height: 3px;
-    background-color: $titleWhiteTheme;
-    transition: all .5s;
-
-    &.menu__button-icon_active {
-      background-color: transparent;
-
-      &::before {
-        transform: rotate(45deg) translate(-3px, -1px);
-      }
-
-      &::after {
-        transform: rotate(-45deg) translate(3px, -5px);
-      }
-    }
-
-    &::before, &::after {
-      content: '';
-      display: block;
-      background-color: $titleWhiteTheme;
-      width: 22px;
-      height: 3px;
-      transition: all .5s;
-    }
-    &::before {
-      transform: translateY(-8px);
-    }
-
-    &::after {
-      transform: translateY(5px);
-    }
+  
+  &:focus {
+    outline: none;
   }
 }
