+++ src/blocks/merckx/merckx.scss
@@ -30,7 +30,7 @@
     &::before {
       content: '';
       display: block;
-      border: 1px solid #e3e3e3;
+      border-top: 1px solid #e3e3e3;
       margin-bottom: 20px;
     }
   }
