+++ src/blocks/radiobox/__button/_hidden/radiobox__button_hidden.scss
@@ -1,8 +1,9 @@
-.radiobox__button_hidden {
+.radiobox__button_hidden.radiobox__button_input {
   position: absolute;
   width: 1px;
   height: 1px;
   overflow: hidden;
+  visibility: hidden;
   clip-path: inset(0 0 0 0);
 
   &:hover + .radiobox__button {
@@ -12,6 +13,27 @@
   &:checked + .radiobox__button {
     font-weight: $semiBold;
     opacity: 1;
+  }
+
+  &:hover:checked + .radiobox__button {
+    opacity: .75;
+  }
+}
+
+.radiobox__button_hidden.radiobox__button_input-mobile {
+  position: absolute;
+  width: 1px;
+  height: 1px;
+  overflow: hidden;
+  visibility: hidden;
+  clip-path: inset(0 0 0 0);
+  
+  &:hover + .radiobox__button {
+    opacity: .6;
+  }
+
+  &:checked + .radiobox__button {
+    background-color: $titleWhiteTheme;
   }
 
   &:hover:checked + .radiobox__button {
