+++ src/blocks/radiobox/__button/_mobile/radiobox__button_mobile.scss
@@ -1,0 +1,8 @@
+.radiobox__button_mobile {
+  display: block;
+  width: 4px;
+  height: 4px;
+  border-radius: 50%;
+  cursor: pointer;
+  background-color: #cacaca;
+}
