--- src/blocks/radiobox/__label/radiobox__label.scss
@@ -1,3 +1,0 @@
-.radiobox__label {
-  
-}
