+++ src/blocks/radiobox/_mobile/radiobox_mobile.scss
@@ -1,0 +1,15 @@
+.radiobox_mobile {
+  display: none;
+
+  @media screen and (max-width: 767px) {
+    display: block;
+    @include flex(row, start, start);
+    max-width: 22px;
+    width: 100%;
+    height: 4px;
+    column-gap: 5px;
+    padding-top: 20px;
+    bottom: 40px;
+    align-self: center;
+  }
+}
