+++ src/blocks/select-wrapper/select-wrapper.scss
@@ -6,8 +6,8 @@
     display: block;
     &::after {
       content: '';
-      width: 10px;
-      height: 5px;
+      width: 12px;
+      height: 7px;
       background-image: url('../images/select-arrow.svg');
       background-size: contain;
       background-repeat: no-repeat;
