+++ src/blocks/selector/__option/selector__option.scss
@@ -1,0 +1,5 @@
+.selector__option {
+  font-size: 11px;
+  line-height: 14px;
+  font-weight: $medium;
+}
