+++ src/blocks/slider/__image-container/slider__image-container.scss
@@ -6,7 +6,6 @@
   left: 0;
   position: relative;
   width: 100%;
-  max-width: 100vw;
   transition: left 1s ease;
 
   @media screen and (max-width: 1023px) {
