+++ src/blocks/slider/_mobile/slider_mobile.scss
@@ -1,0 +1,5 @@
+.slider_mobile {
+  @media screen and (max-width: 767px) {
+    max-width: 284px;
+  }
+}
